Manejo de datasets. Funciones de extracción de características
Todo lo relacionado al manejo de arff se derivó al python weka wrapper, casi todas las funciones de arff manager se remplazaron por su equivalente utilizando dicha librería.

Las funciones de extraccion de características de video se adaptaron para utilizar las características extraídas en otra función, al ejecutar un experimento genera una subcarpeta con las características concatenadas y adaptadas a este. No las recalcula cada vez que se ejecutan.

Cambio en el nombre del archivo que contiene los métodos de extracción de características.

Demás cambios menores
